requirements.txt
spacy==2.3.2
wtforms==2.3.3
celery==4.4.7
markupsafe==2.0.1
Flask-OpenID==1.3.0
apache-airflow==1.10.15
autopep8==1.4.4
cryptography==2.9.2
enum-compat==0.0.3
enum34==1.1.6
flake8==3.7.7
Flask-Migrate==2.5.3
flower==0.9.4
futures==3.1.1
gensim==3.7.3
ipaddress==1.0.22
jupyter==1.0.0
many-stop-words==0.2.2
nltk==3.4.1
pathlib2==2.3.3
pathtools==0.1.2
pep8==1.7.1
pipdeptree==2.1.0
pipreqs==0.4.10
psycopg2-binary==2.8.5
pyLDAvis==2.1.2
pyshp==2.1.0
PySocks==1.7.0
python-dotenv==0.13.0
redis==3.5.2
requests-oauthlib==1.2.0
Rx==1.6.1
s3transfer==0.2.0
scandir==1.10.0
sklearn==0.0
streamlit==0.89.0
wordcloud==1.8.1
tweepy==3.7.0
tweet-preprocessor==0.6.0
Flask-Bcrypt==0.7.1
python3-openid==3.2.0
Flask-Script==2.0.6
